// src/types/Reader.ts
// Define the Reader type
export type Reader = {
    _id: string;
    first_name: string;
    last_name: string;
    email: string;
    age: number;
    address: string;
    phone: string;
    join_date: string;
};

export type ReaderFormData = {
    first_name: string;
    last_name: string;
    email: string;
    age: number;
    address: string;
    phone: string;
    join_date: string;

}